[batch] refactor: extract billing query functions into billing_reporting module
